requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
requests
beautifulsoup4
sentence-transformers
faiss-cpu